refactor: :recycle: Nest the sample utils to make it more comprehensive
